Add Medication

+++ Drone/Models/DronesModel.cs
@@ -8,16 +8,17 @@ namespace DronesLoad.Models
     {
         public int Id { get; set; }
 		[Required]
+        
 		[StringLength(maximumLength: 100, ErrorMessage = "Serial Number is required with length less than 100")]
 		public string SerialNumber { get; set; } = null!;
         public int? ModelId { get; set; }
         public int? StateId { get; set; }
 		[Required]
 		[Range(0, 500)]
-		public int? WeightLimit { get; set; }
+		public double? WeightLimit { get; set; }
 		[Required]
-		[Range(0,100)]
-        public decimal? BatteryCapacity { get; set; }
+		[Range(0,100.00)]
+        public double? BatteryCapacity { get; set; }
 
         public virtual DModelWeight? DModel { get; set; }
         public virtual DroneStatesModel? State { get; set; }
